+++ requirements.txt
@@ -14,3 +14,4 @@ ttkthemes>=3.0.0
 send2trash>=1.8.0,<2.0.0
 psutil>=5.8.0,<6.0.0
 tree-sitter==0.20.0
+PyYAML==6.0
