+++ tests/test_highlight_plugin.py
@@ -1,3 +1,7 @@
+import subprocess
+import sys
+
+
 from pygments.lexers import PythonLexer
 
 
@@ -50,3 +54,33 @@ def test_last_line_bug(filetab, tmp_path):
     filetab.textwidget.insert("1.0", "# This is a comment")
     filetab.update()
     assert filetab.textwidget.tag_names("1.5") == ("Token.Comment.Single",)
+
+
+# I currently don't think the tree-sitter highlighter needs lots of tests.
+# If it doesn't work, it's usually quite obvious after using it a while.
+#
+# That said, it comes with a couple fragile things:
+#    - Unzipping and loading the correct binary file on each platform
+#    - The dumping script, used only when configuring the plugin
+#
+# These need testing, and this test conveniently tests them both.
+def test_dumping_hello_world_program(tmp_path):
+    (tmp_path / "hello.py").write_text("print('hello')")
+    output = subprocess.check_output(
+        [sys.executable, "scripts/tree-sitter-dump.py", "python", str(tmp_path / "hello.py")],
+        text=True,
+    )
+
+    expected_output = """
+type=module text="print('hello')"
+  type=expression_statement text="print('hello')"
+    type=call text="print('hello')"
+      type=identifier text='print'
+      type=argument_list text="('hello')"
+        type=( text='('
+        type=string text="'hello'"
+          type=" text="'"
+          type=" text="'"
+        type=) text=')'
+    """
+    assert output.strip() == expected_output.strip()
